mypy fixes
