+++ ptychodus/controller/ptychopy.py
@@ -4,7 +4,7 @@ from decimal import Decimal
 from PyQt5.QtWidgets import QWidget
 
 from ..api.observer import Observable, Observer
-from ..model import PtychoPyReconstructorLibrary, PtychoPyPresenter
+from ..model.ptychopy import PtychoPyReconstructorLibrary, PtychoPyPresenter
 from ..view import PtychoPyParametersView, PtychoPyBasicView, PtychoPyAdvancedView
 from .reconstructor import ReconstructorViewControllerFactory
